songs_api/songs_api/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
mysql-connector-python==8.0.20
mysqlclient==1.4.6